requirements.txt
# Q.Invoice - Python Dependencies
# Generated: January 2026

# Core Framework
streamlit==1.32.0

# OCR & Image Processing
paddleocr==2.7.3
paddlepaddle==2.6.0
opencv-python==4.9.0.80
pillow==10.2.0

# PDF Processing
PyMuPDF==1.23.22
pdf2image==1.17.0

# LLM & AI
openai==1.12.0
anthropic==0.18.1

# Database
sqlalchemy==2.0.27
alembic==1.13.1

# Data Processing
pandas==2.2.0
openpyxl==3.1.2
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
pydantic==2.6.1
pydantic-settings==2.1.0

# Export
xlsxwriter==3.1.9

# Testing (optional, for development)
pytest==8.0.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0